refactor: replace SceneFactory with functional scene builder
